js text de la home page, map changed'
