+++ app/views/shared/_navbar.html.erb
@@ -1,6 +1,6 @@
 <div class="navbar navbar-expand-sm sticky-top px-5">
   <div class= "d-flex justify-content-around align-items-center">
-    <%= link_to "/", class: "navbar-brand" do %>
+    <%= link_to "/", class: "navbar-brand", remote: false do %>
       <%= image_tag "timecapsule-incaps.gif", width: 500, height: 95 %>
     <% end %>
     <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
